Creates stamina regeneration at x rate if it is not used after y seconds.

+++ Assets/Scripts/Enemy/Action_Attack.cs
@@ -2,7 +2,7 @@
 
 [CreateAssetMenu (menuName = "AI/Actions/Enemy_Attack")]
 public class Action_Attack : Action {
-	[SerializeField] int staminaRequired = 2;
+	[SerializeField] float staminaRequired = 2f;
 	[SerializeField] float attackDelay = 5f;
 	
 	float timer = 0f;
